Basic functions for Collision
The basic isCollidin, is Inside and rangeCheck functions have been placed. These functions will be moved into their own file to allow the player and the gameobjects to inherit. This is just a basis for now.

The is Colliding Function will check for collision along the bounding boxs.

The is Inside Function will check if another collision box is inside the curent object.

The Range Check Function will show the range of the object outside its given bounding box.