Added a punch hit box.
Created a new object called punch_box so that the collider vector in the collision manager stays synchronised with the vector of objects.
Before this I tried adding two colliders to the one player object but this caused every 2 objects two have 4 colliders so my collison code bugged and only scanned the first two colliders being
the first player and first punch box.

Working on the box's ability to follow the player and switch sides when the player does.