Colliders now use pointers instead of passing in locations via function parameters.
This allows all the colliders to be updated specific to their players location using the WASD update system for player movement.
With this in place the punch box now switches position based on the last direction the player was facing.

Should be able to implement a punch on trigger with the punch box, key presses and addForce functionality already built in.