Collision now sets the position of the players based on overlap not physics, punch box position fixed.
Players now use the difference between box origins / 100 to give  less rebound on collision (seems more realiistic). The x velocityis also set to zero.

Punch boxes are now positioned based on an offset that changes when A or D are pressed rather than the position changing itself. This means the punch colliders can be updated constantly rather than just when holding down A or D which caused logic errors.