Added new cursor, uses gamepad input to move
Cursor inherits from UIObject and uses composition for UISprite. 

The sprite is rendered inside the Cursors render function, so when instanciated the cursor can be used rather than the sprite inside.

Using the built in DirectX functionality I can create floats based on thumbstick axis and set the position
of the cursor to be incremented based on stick axis * speed.